requirements.txt
attrs==23.2.0
aws-cdk-lib==2.149.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
cattrs==23.2.3
constructs==10.3.0
exceptiongroup==1.2.2
importlib_resources==6.4.0
jsii==1.101.0
prettytable==3.10.2
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
wcwidth==0.2.13
zipp==3.19.2
